new file for DA and f
